==> scripts/02_energy_movement.Rmd <==
---
editor_options: 
  chunk_output_type: console
---

# Mean Energy and Movement Across Scenarios

```{r}
# load libraries
library(data.table)

# plotting
library(ggplot2)
library(glue)
library(patchwork)
library(biscale)
```

## Load energy data

```{r}
load("data/data_energy_move.Rds")
```

Unlist and make larger frame.

```{r}
# unlist
data_energy = data_energy[, unlist(dt, recursive = FALSE), 
            by = setdiff(names(data_energy), "dt")]
```

Select every Nth generation for plotting.

```{r}
data_energy = data_energy[gen %% 10 == 0, ]
```

## Select frequency

```{r}
data_energy = data_energy[freqRes == 1, ]
```


## Plotting

### Split data by variable and scenario

```{r}
# by variable
data_energy = split(data_energy, by = "variable")

# by scenario
data_energy = lapply(data_energy, function(df) {
  split(df, by = "scenario")
})
```

### Plot energy

```{r}
plot_energy = lapply(data_energy[["energy"]], function(df) {
  
  scenario = unique(df$scenario)
  scenario = dplyr::case_when(
    scenario == 0 ~ "no pathogen",
    scenario == 1 ~ "ancestral pathogen",
    scenario == 2 ~ "spillover pathogen",
    TRUE ~ NA_character_
  )
  
  genmax = max(df$gen)
  
  p = df |>
    ggplot()+
    geom_ribbon(
      aes(
        x = gen,  
        ymin = mean - sd,
        ymax = mean + sd,
        group = replicate
      ),
      fill = alpha("grey", alpha = 0.1)
    )+
    geom_line(
      aes(
        gen, mean, 
        group = replicate
      ),
      show.legend = F
    )+
    coord_cartesian(
      expand = F,
      ylim = c(0, 20)
    )+
    facet_grid(
      pTransmit ~ costInfection, 
      scales = "fixed",
      labeller = label_both,
      switch = "y"
    )+
    theme_test()+
    labs(
      x = "Gen.",
      y = "Mean per-capita energy",
      title = glue::glue(
        "scenario: {scenario}"
      )
    )
  
  if (scenario == "spillover pathogen") {
    p = p + geom_vline(
      xintercept = genmax * (2/3),
      col = "red",
      lty = 2
    )
  }
  p
})
```


```{r}
plot = wrap_plots(
  plot_energy
)

ggsave(
  plot = plot,
  filename = "figures/fig_energy_scenario.png"
)
```

### Plot movement

```{r}
plot_movement = lapply(data_energy[["moved"]], function(df) {
  
  scenario = unique(df$scenario)
  scenario = dplyr::case_when(
    scenario == 0 ~ "no pathogen",
    scenario == 1 ~ "ancestral pathogen",
    scenario == 2 ~ "spillover pathogen",
    TRUE ~ NA_character_
  )
  
  genmax = max(df$gen)
  
  p = df |>
    ggplot()+
    geom_ribbon(
      aes(
        x = gen,  
        ymin = mean - sd,
        ymax = mean + sd,
        group = replicate
      ),
      fill = alpha("grey", alpha = 0.1)
    )+
    geom_line(
      aes(
        gen, mean, 
        group = replicate
      ),
      col = "steelblue",
      show.legend = F
    )+
    # coord_cartesian(
    #   expand = F,
    #   ylim = c(5, 20)
    # )+
    facet_grid(
      pTransmit ~ costInfection, 
      scales = "fixed",
      labeller = label_both,
      switch = "y"
    )+
    theme_test()+
    labs(
      x = "Gen.",
      y = "Mean per-capita movement",
      title = glue::glue(
        "scenario: {scenario}"
      )
    )
  
  if (scenario == "spillover pathogen") {
    p = p + geom_vline(
      xintercept = genmax * (2/3),
      col = "red",
      lty = 2
    )
  }
  p
})
```


```{r}
plot = wrap_plots(
  plot_movement
)

ggsave(
  plot = plot,
  filename = "figures/fig_movement_scenario.png"
)
```
